requirements.txt
redis
aiohttp
starlette
uvicorn
pyyaml